rollback benchmarked HRI message to rai_bench package
